tối ưu tách components dễ quản lý auth
