--- product-shopthethao/ShopeTheThao_frontend_react/src/components/User/Auth/ChangePassword/ChangePasswordForm.jsx
@@ -1,95 +1,0 @@
-import React, { useState } from 'react';
-import { Form, Input, Button, message } from 'antd';
-import { LockOutlined } from '@ant-design/icons';
-
-import { useNavigate } from 'react-router-dom';
-import authApi from 'api/Admin/Auth/auth';
-
-const ChangePasswordForm = () => {
-  const [loading, setLoading] = useState(false);
-  const navigate = useNavigate();
-  const [form] = Form.useForm();
-
-  const onFinish = async (values) => {
-    setLoading(true);
-    try {
-      await authApi.changePasswordNew({
-        oldPassword: values.oldPassword,
-        newPassword: values.newPassword
-      });
-      message.success('Đổi mật khẩu thành công!');
-      form.resetFields();
-      navigate('/profile');
-    } catch (error) {
-      message.error('Đổi mật khẩu thất bại: ' + (error.response?.data?.message || 'Có lỗi xảy ra'));
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  return (
-    <div className="change-password-container">
-      <h2>Đổi Mật Khẩu</h2>
-      <Form
-        form={form}
-        name="changePassword"
-        onFinish={onFinish}
-        layout="vertical"
-      >
-        <Form.Item
-          name="oldPassword"
-          rules={[
-            { required: true, message: 'Vui lòng nhập mật khẩu hiện tại!' }
-          ]}
-        >
-          <Input.Password 
-            prefix={<LockOutlined />} 
-            placeholder="Mật khẩu hiện tại" 
-          />
-        </Form.Item>
-
-        <Form.Item
-          name="newPassword"
-          rules={[
-            { required: true, message: 'Vui lòng nhập mật khẩu mới!' },
-            { min: 6, message: 'Mật khẩu phải có ít nhất 6 ký tự!' }
-          ]}
-        >
-          <Input.Password 
-            prefix={<LockOutlined />} 
-            placeholder="Mật khẩu mới" 
-          />
-        </Form.Item>
-
-        <Form.Item
-          name="confirmPassword"
-          dependencies={['newPassword']}
-          rules={[
-            { required: true, message: 'Vui lòng xác nhận mật khẩu!' },
-            ({ getFieldValue }) => ({
-              validator(_, value) {
-                if (!value || getFieldValue('newPassword') === value) {
-                  return Promise.resolve();
-                }
-                return Promise.reject(new Error('Mật khẩu xác nhận không khớp!'));
-              },
-            }),
-          ]}
-        >
-          <Input.Password 
-            prefix={<LockOutlined />} 
-            placeholder="Xác nhận mật khẩu mới" 
-          />
-        </Form.Item>
-
-        <Form.Item>
-          <Button type="primary" htmlType="submit" loading={loading} block>
-            Đổi Mật Khẩu
-          </Button>
-        </Form.Item>
-      </Form>
-    </div>
-  );
-};
-
-export default ChangePasswordForm;
