+++ product-shopthethao/ShopeTheThao_frontend_react/src/components/User/Auth/EmailVerification/emailVerification.scss
@@ -1,0 +1,170 @@
+@import '../styles/variables';
+
+.email-verification-container {
+  min-height: 100vh;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  background: linear-gradient(135deg, #4F46E5, #7C3AED);
+  padding: 1rem;
+  
+  .verification-card {
+    background: white;
+    border-radius: 1rem;
+    padding: 3rem 2.5rem;
+    width: 100%;
+    max-width: 500px;
+    box-shadow: 0 10px 25px rgba(0, 0, 0, 0.1);
+    text-align: center;
+    
+    .verification-header {
+      margin-bottom: 2rem;
+      
+      .email-icon {
+        width: 80px;
+        height: 80px;
+        border-radius: 50%;
+        background: rgba($primary-color, 0.1);
+        color: $primary-color;
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        margin: 0 auto 1.5rem;
+      }
+      
+      h2 {
+        font-size: 1.75rem;
+        font-weight: 700;
+        color: $text-primary;
+        margin-bottom: 1rem;
+      }
+      
+      p {
+        color: $text-secondary;
+        font-size: 0.95rem;
+        line-height: 1.5;
+      }
+    }
+    
+    .verification-form {
+      .otp-input-wrapper {
+        margin-bottom: 1.5rem;
+        
+        input {
+          text-align: center;
+          letter-spacing: 0.2em;
+          font-size: 1.5rem;
+        }
+      }
+      
+      .verify-button {
+        width: 100%;
+        padding: 1rem;
+        background: $primary-color;
+        color: white;
+        border: none;
+        border-radius: 0.5rem;
+        font-weight: 600;
+        cursor: pointer;
+        transition: all 0.2s ease;
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        gap: 0.5rem;
+        
+        .button-icon {
+          transition: transform 0.3s ease;
+        }
+        
+        &:hover:not(:disabled) {
+          background: darken($primary-color, 5%);
+          transform: translateY(-2px);
+          
+          .button-icon {
+            transform: scale(1.2);
+          }
+        }
+        
+        &:disabled {
+          opacity: 0.7;
+          cursor: not-allowed;
+        }
+        
+        &.loading {
+          position: relative;
+          padding-left: 2.5rem;
+          
+          &::before {
+            content: '';
+            position: absolute;
+            left: 1rem;
+            top: 50%;
+            transform: translateY(-50%);
+            width: 1.2rem;
+            height: 1.2rem;
+            border: 2px solid rgba(255, 255, 255, 0.3);
+            border-radius: 50%;
+            border-top-color: white;
+            animation: spin 1s linear infinite;
+          }
+        }
+      }
+      
+      .resend-wrapper {
+        margin-top: 1.5rem;
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        gap: 0.5rem;
+        
+        .resend-button {
+          background: transparent;
+          border: none;
+          color: $primary-color;
+          font-weight: 500;
+          cursor: pointer;
+          padding: 0.5rem;
+          display: flex;
+          align-items: center;
+          gap: 0.3rem;
+          transition: all 0.2s ease;
+          
+          .button-icon {
+            transition: transform 0.3s ease;
+          }
+          
+          &:hover:not(.disabled) {
+            color: darken($primary-color, 10%);
+            
+            .button-icon {
+              transform: rotate(180deg);
+            }
+          }
+          
+          &.disabled {
+            color: $text-secondary;
+            cursor: not-allowed;
+            opacity: 0.7;
+          }
+        }
+        
+        .countdown {
+          color: $text-secondary;
+          font-size: 0.85rem;
+        }
+      }
+    }
+  }
+}
+
+@keyframes spin {
+  to { transform: rotate(360deg) translateY(-50%); }
+}
+
+@media (max-width: 640px) {
+  .email-verification-container {
+    .verification-card {
+      padding: 2rem 1.5rem;
+    }
+  }
+}
