+++ product-shopthethao/ShopeTheThao_frontend_react/src/components/User/Auth/Login/LoginForm.jsx
@@ -1,32 +1,19 @@
 import React, { useState, useEffect } from "react";
 import { useNavigate, Link, useLocation } from "react-router-dom";
-import {
-  FiLock,
-  FiEye,
-  FiEyeOff,
-  FiUser,
-  FiMail,
-  FiArrowRight,
-  FiPhone,
-  FiInfo,
-} from "react-icons/fi";
-import {
-  FaFacebookF,
-  FaGoogle,
-  FaTwitter,
-  FaGithub,
-  FaMars,
-  FaVenus,
-  FaUserAlt,
-} from "react-icons/fa";
 import { motion, AnimatePresence } from "framer-motion";
-import { Tabs, message, Divider, Radio } from "antd";
+import { Tabs, message, Divider } from "antd";
+import { FiLock, FiEye, FiEyeOff, FiUser, FiMail, FiArrowRight } from "react-icons/fi";
+import { FaFacebookF, FaGoogle, FaTwitter, FaGithub } from "react-icons/fa";
 import "./loginForm.scss";
 import img from "assets/Img";
 import { useAuth } from "hooks/useAuth";
 import { getRedirectPath, getLoginMessage } from "utils/roleManager";
 import { validateId, validatePassword } from "../Custom";
-import authApi from "api/Admin/Auth/auth";
+
+// Fix import paths to directly import from the FormFields file
+import { InputField, CustomCheckbox, SocialButton } from "../Common/FormFields";
+import ForgotPasswordForm from "../ForgotPassword/ForgotPasswordForm";
+import RegisterForm from "../Register/RegisterForm";
 
 const LoginForm = () => {
   const navigate = useNavigate();
@@ -125,12 +112,10 @@ const LoginForm = () => {
   const handleForgotPassword = async (e) => {
     e.preventDefault();
     setShowForgotPassword(true);
-    // Now the rest of the functionality is handled within the ForgotPasswordForm component
   };
 
   const handleSocialLogin = (provider) => {
     message.info(`Đăng nhập bằng ${provider} đang được phát triển!`);
-    // Here you would implement OAuth login for the provider
   };
 
   useEffect(() => {
@@ -231,9 +216,7 @@ const LoginForm = () => {
               exit={{ opacity: 0, y: -10 }}
               transition={{ duration: 0.3 }}
             >
-              {/* ...existing login form... */}
               <form onSubmit={handleSubmit} className="login-form">
-                {/* ...existing login form fields... */}
                 <InputField
                   icon={<FiUser />}
                   type="text"
@@ -332,7 +315,7 @@ const LoginForm = () => {
             exit={{ opacity: 0, y: -10 }}
             transition={{ duration: 0.3 }}
           >
-            <EnhancedRegisterForm
+            <RegisterForm
               onLoginClick={() => handleTabChange("login")}
             />
           </motion.div>
@@ -398,929 +381,5 @@ const LoginForm = () => {
     </div>
   );
 };
-
-// Enhanced Register Form Component
-const EnhancedRegisterForm = ({ onLoginClick }) => {
-  const [formData, setFormData] = useState({
-    id: "",
-    phone: "",
-    fullname: "",
-    email: "",
-    password: "",
-    confirmPassword: "",
-    gender: "M", // M: Male, F: Female, O: Other
-    role: ["USER"],
-  });
-  const [errors, setErrors] = useState({
-    id: "",
-    phone: "",
-    fullname: "",
-    email: "",
-    password: "",
-    confirmPassword: "",
-  });
-  const [loading, setLoading] = useState(false);
-  const [showPassword, setShowPassword] = useState(false);
-  const [showConfirmPassword, setShowConfirmPassword] = useState(false);
-  const [step, setStep] = useState(1);
-  const [registrationComplete, setRegistrationComplete] = useState(false);
-  const [passwordStrength, setPasswordStrength] = useState({
-    score: 0,
-    label: "",
-  });
-  const navigate = useNavigate();
-
-  const validateField = (name, value) => {
-    switch (name) {
-      case "id":
-        if (!value.trim()) return "Tên tài khoản không được để trống";
-        if (value.length < 4) return "Tên tài khoản phải có ít nhất 4 ký tự";
-        if (!/^[a-zA-Z0-9_]+$/.test(value))
-          return "Tên tài khoản chỉ chứa chữ cái, số và dấu gạch dưới";
-        return "";
-      case "phone":
-        if (!value.trim()) return "Số điện thoại không được để trống";
-        if (!/^[0-9]{10}$/.test(value.replace(/\s/g, "")))
-          return "Số điện thoại phải có 10 chữ số";
-        return "";
-      case "fullname":
-        if (!value.trim()) return "Họ và tên không được để trống";
-        if (value.trim().length < 2) return "Họ và tên quá ngắn";
-        return "";
-      case "email":
-        if (!value.trim()) return "Email không được để trống";
-        if (!/^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(value))
-          return "Email không hợp lệ";
-        return "";
-      case "password":
-        if (!value) return "Mật khẩu không được để trống";
-        if (value.length < 6) return "Mật khẩu phải có ít nhất 6 ký tự";
-        if (!/[A-Za-z]/.test(value) || !/[0-9]/.test(value))
-          return "Mật khẩu phải chứa cả chữ và số";
-        return "";
-      case "confirmPassword":
-        if (!value) return "Vui lòng xác nhận mật khẩu";
-        if (value !== formData.password) return "Mật khẩu xác nhận không khớp";
-        return "";
-      default:
-        return "";
-    }
-  };
-
-  const handleChange = (e) => {
-    const { name, value } = e.target;
-    setFormData((prev) => ({
-      ...prev,
-      [name]: value,
-    }));
-
-    // Validate field on change
-    setErrors((prev) => ({
-      ...prev,
-      [name]: validateField(name, value),
-    }));
-
-    // Calculate password strength if password field changes
-    if (name === "password") {
-      calculatePasswordStrength(value);
-    }
-  };
-
-  const calculatePasswordStrength = (password) => {
-    if (!password) {
-      setPasswordStrength({ score: 0, label: "" });
-      return;
-    }
-
-    let score = 0;
-
-    // Length check
-    if (password.length >= 6) score += 1;
-    if (password.length >= 10) score += 1;
-
-    // Character variety check
-    if (/[A-Z]/.test(password)) score += 1;
-    if (/[0-9]/.test(password)) score += 1;
-    if (/[^A-Za-z0-9]/.test(password)) score += 1;
-
-    let label = "";
-    if (score <= 2) label = "weak";
-    else if (score <= 4) label = "medium";
-    else label = "strong";
-
-    setPasswordStrength({ score, label });
-  };
-
-  const handleGenderChange = (e) => {
-    setFormData((prev) => ({
-      ...prev,
-      gender: e.target.value,
-    }));
-  };
-
-  const validateStep = (currentStep) => {
-    let isValid = true;
-    const newErrors = { ...errors };
-
-    if (currentStep === 1) {
-      // Validate Step 1 fields
-      const fields = ["id", "phone", "fullname"];
-      fields.forEach((field) => {
-        const error = validateField(field, formData[field]);
-        newErrors[field] = error;
-        if (error) isValid = false;
-      });
-    } else if (currentStep === 2) {
-      // Validate Step 2 fields
-      const fields = ["email", "password", "confirmPassword"];
-      fields.forEach((field) => {
-        const error = validateField(field, formData[field]);
-        newErrors[field] = error;
-        if (error) isValid = false;
-      });
-    }
-
-    setErrors(newErrors);
-    return isValid;
-  };
-
-  const nextStep = () => {
-    if (validateStep(step)) {
-      // Add smooth animation between steps
-      const content = document.querySelector(".step-content");
-      if (content) {
-        content.classList.add("fade-out");
-        setTimeout(() => {
-          setStep(step + 1);
-          content.classList.remove("fade-out");
-        }, 300);
-      } else {
-        setStep(step + 1);
-      }
-    } else {
-      // Shake animation for validation errors
-      const form = document.querySelector(".register-form");
-      form.classList.add("shake");
-      setTimeout(() => form.classList.remove("shake"), 500);
-    }
-  };
-
-  const prevStep = () => {
-    setStep(step - 1);
-  };
-
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-    if (!validateStep(step)) {
-      message.error("Vui lòng điền đầy đủ thông tin!");
-      return;
-    }
-
-    setLoading(true);
-    try {
-      // Create the registration data in the correct format
-      const registrationData = {
-        id: formData.id,
-        phone: formData.phone,
-        fullname: formData.fullname,
-        email: formData.email,
-        password: formData.password,
-        gender: formData.gender, // Send M/F/O directly
-        role: ["USER"],
-      };
-
-      // Remove confirmPassword as it's not needed in the API
-      delete registrationData.confirmPassword;
-
-      // Call the API
-      await authApi.signup(registrationData);
-      message.success(
-        "Đăng ký thành công! Kiểm tra email nhập mã để xác nhận tài khoản."
-      );
-      navigate("/v1/auth/otp", { state: { id: formData.id } });
-    } catch (error) {
-      const errorMessage =
-        error.response?.data?.message || "Đăng ký thất bại. Vui lòng thử lại!";
-      message.error(errorMessage);
-      console.error("Registration error:", error);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const renderStepIndicator = () => (
-    <div className="step-indicator">
-      <div
-        className={`step ${step >= 1 ? "active" : ""} ${
-          step > 1 ? "completed" : ""
-        }`}
-      >
-        <div className="step-number">1</div>
-        <div className="step-title">Thông tin cá nhân</div>
-      </div>
-      <div className={`step-connector ${step > 1 ? "half" : ""}`}></div>
-      <div className={`step ${step >= 2 ? "active" : ""}`}>
-        <div className="step-number">2</div>
-        <div className="step-title">Thông tin tài khoản</div>
-      </div>
-    </div>
-  );
-
-  const renderPasswordStrength = () => (
-    <div className="password-strength">
-      <span className={`strength-text ${passwordStrength.label}`}>
-        {passwordStrength.label === "weak" && "Yếu"}
-        {passwordStrength.label === "medium" && "Trung bình"}
-        {passwordStrength.label === "strong" && "Mạnh"}
-      </span>
-      <div className="strength-bars">
-        <div
-          className={`bar ${
-            passwordStrength.score >= 1
-              ? `active ${passwordStrength.label}`
-              : ""
-          }`}
-        ></div>
-        <div
-          className={`bar ${
-            passwordStrength.score >= 2
-              ? `active ${passwordStrength.label}`
-              : ""
-          }`}
-        ></div>
-        <div
-          className={`bar ${
-            passwordStrength.score >= 3
-              ? `active ${passwordStrength.label}`
-              : ""
-          }`}
-        ></div>
-        <div
-          className={`bar ${
-            passwordStrength.score >= 4
-              ? `active ${passwordStrength.label}`
-              : ""
-          }`}
-        ></div>
-        <div
-          className={`bar ${
-            passwordStrength.score >= 5
-              ? `active ${passwordStrength.label}`
-              : ""
-          }`}
-        ></div>
-      </div>
-    </div>
-  );
-
-  if (registrationComplete) {
-    return (
-      <div className="register-success">
-        <motion.div
-          className="success-icon"
-          initial={{ scale: 0 }}
-          animate={{ scale: 1 }}
-          transition={{ type: "spring", stiffness: 260, damping: 20 }}
-        >
-          <svg
-            viewBox="0 0 24 24"
-            fill="none"
-            xmlns="http://www.w3.org/2000/svg"
-          >
-            <circle
-              cx="12"
-              cy="12"
-              r="10"
-              stroke="currentColor"
-              strokeWidth="1.5"
-            />
-            <path
-              d="M8 12L11 15L16 9"
-              stroke="currentColor"
-              strokeWidth="1.5"
-              strokeLinecap="round"
-              strokeLinejoin="round"
-            />
-          </svg>
-        </motion.div>
-        <h3>Đăng ký thành công!</h3>
-        <p>Tài khoản của bạn đã được tạo. Vui lòng đăng nhập để tiếp tục.</p>
-        <motion.button
-          className="submit-button login-button"
-          onClick={onLoginClick}
-          whileTap={{ scale: 0.98 }}
-        >
-          Đăng nhập ngay
-        </motion.button>
-      </div>
-    );
-  }
-
-  return (
-    <motion.form
-      onSubmit={handleSubmit}
-      className={`register-form ${loading ? "submitting" : ""}`}
-      initial={{ opacity: 0 }}
-      animate={{ opacity: 1 }}
-      transition={{ duration: 0.3 }}
-    >
-      {renderStepIndicator()}
-      <AnimatePresence mode="wait">
-        {step === 1 ? (
-          <motion.div
-            key="step1"
-            initial={{ opacity: 0, x: -20 }}
-            animate={{ opacity: 1, x: 0 }}
-            exit={{ opacity: 0, x: 20 }}
-            transition={{ duration: 0.3 }}
-            className="step-content animated-input"
-          >
-            <div className="form-group-wrapper">
-              <InputField
-                icon={<FaUserAlt />}
-                type="text"
-                name="id"
-                placeholder="Tên tài khoản"
-                value={formData.id}
-                onChange={handleChange}
-                error={errors.id}
-                maxLength={30}
-                toolTip="Tên tài khoản dùng để đăng nhập (ví dụ: johndoe123)"
-              />
-            </div>
-            <div className="form-group-wrapper">
-              <InputField
-                icon={<FiPhone />}
-                type="tel"
-                name="phone"
-                placeholder="Số điện thoại"
-                value={formData.phone}
-                onChange={handleChange}
-                error={errors.phone}
-                maxLength={10}
-              />
-            </div>
-            <div className="form-group-wrapper">
-              <InputField
-                icon={<FiUser />}
-                type="text"
-                name="fullname"
-                placeholder="Họ và tên đầy đủ"
-                value={formData.fullname}
-                onChange={handleChange}
-                error={errors.fullname}
-              />
-            </div>
-            <div className="form-group">
-              <label className="input-label">Giới tính</label>
-              <div className="gender-select">
-                <Radio.Group
-                  onChange={handleGenderChange}
-                  value={formData.gender}
-                >
-                  <Radio value="M" className="gender-option">
-                    <FaMars className="gender-icon male" /> Nam
-                  </Radio>
-                  <Radio value="F" className="gender-option">
-                    <FaVenus className="gender-icon female" /> Nữ
-                  </Radio>
-                  <Radio value="O" className="gender-option">
-                    <FiUser className="gender-icon other" /> Khác
-                  </Radio>
-                </Radio.Group>
-              </div>
-            </div>
-            <motion.button
-              type="button"
-              className="submit-button"
-              onClick={nextStep}
-              whileTap={{ scale: 0.98 }}
-              whileHover={{ boxShadow: "0 5px 15px rgba(0, 0, 0, 0.1)" }}
-            >
-              Tiếp theo
-            </motion.button>
-          </motion.div>
-        ) : (
-          <motion.div
-            key="step2"
-            initial={{ opacity: 0, x: 20 }}
-            animate={{ opacity: 1, x: 0 }}
-            exit={{ opacity: 0, x: -20 }}
-            transition={{ duration: 0.3 }}
-            className="step-content animated-input"
-          >
-            <div className="form-group-wrapper">
-              <EmailField
-                icon={<FiMail />}
-                name="email"
-                placeholder="Email"
-                value={formData.email}
-                onChange={handleChange}
-                error={errors.email}
-              />
-            </div>
-            <div className="form-group-wrapper">
-              <InputField
-                icon={<FiLock />}
-                type={showPassword ? "text" : "password"}
-                name="password"
-                placeholder="Mật khẩu"
-                value={formData.password}
-                onChange={handleChange}
-                error={errors.password}
-                endIcon={
-                  <button
-                    type="button"
-                    className="visibility-toggle"
-                    onClick={() => setShowPassword(!showPassword)}
-                  >
-                    {showPassword ? <FiEyeOff /> : <FiEye />}
-                  </button>
-                }
-                requirements="Tối thiểu 6 ký tự, bao gồm chữ và số"
-              />
-            </div>
-            {formData.password && renderPasswordStrength()}
-            <div className="form-group-wrapper">
-              <InputField
-                icon={<FiLock />}
-                type={showConfirmPassword ? "text" : "password"}
-                name="confirmPassword"
-                placeholder="Xác nhận mật khẩu"
-                value={formData.confirmPassword}
-                onChange={handleChange}
-                error={errors.confirmPassword}
-                endIcon={
-                  <button
-                    type="button"
-                    className="visibility-toggle"
-                    onClick={() => setShowConfirmPassword(!showConfirmPassword)}
-                  >
-                    {showConfirmPassword ? <FiEyeOff /> : <FiEye />}
-                  </button>
-                }
-              />
-            </div>
-            <div className="button-group">
-              <motion.button
-                type="button"
-                className="back-button"
-                onClick={prevStep}
-                // whileTap={{ scale: 0.98 }}
-                disabled={loading}
-              >
-                Quay lại
-              </motion.button>
-              <motion.button
-                type="submit"
-                className="submit-button"
-                disabled={loading}
-                whileTap={{ scale: 0.98 }}
-                whileHover={{ boxShadow: "0 5px 15px rgba(0, 0, 0, 0.1)" }}
-              >
-                {loading ? <>ĐANG XỬ LÝ...</> : "ĐĂNG KÝ"}
-              </motion.button>
-            </div>
-          </motion.div>
-        )}
-      </AnimatePresence>
-      <div className="form-footer">
-        <p>Đã có tài khoản?</p>
-        <button 
-          type="button"
-          className="switch-auth-mode"
-          onClick={onLoginClick}
-        >
-          Đăng nhập ngay
-          <FiArrowRight className="arrow-icon" />
-        </button>
-      </div>
-    </motion.form>
-  );
-};
-
-// Optimized helper components
-const InputField = ({
-  icon,
-  endIcon,
-  error,
-  toolTip,
-  requirements,
-  enableEmailSuggestions = false,
-  ...props
-}) => {
-  const [inputValue, setInputValue] = useState(props.value || '');
-  
-  useEffect(() => {
-    setInputValue(props.value || '');
-  }, [props.value]);
-  
-  const handleEmailInput = (e) => {
-    let value = e.target.value;
-    
-    // Remove any existing email domain to prevent duplications
-    if (enableEmailSuggestions && !value.endsWith('@gmail.com') && !value.includes('@gmail.com')) {
-      // If the user just started typing and hasn't entered @ yet
-      if (!value.includes('@') && value.length > 0) {
-        // Auto-append @gmail.com to the end of the input
-        value = value + '@gmail.com';
-      }
-    }
-    
-    setInputValue(value);
-    
-    // Create a synthetic event object with the modified value
-    const syntheticEvent = {
-      target: {
-        name: e.target.name,
-        value: value,
-      }
-    };
-    
-    props.onChange(syntheticEvent);
-  };
-  
-  const handleKeyDown = (e) => {
-    // If user presses @ manually, replace with @gmail.com
-    if (e.key === '@' && enableEmailSuggestions) {
-      e.preventDefault();
-      const cursorPosition = e.target.selectionStart;
-      const valueBeforeCursor = e.target.value.substring(0, cursorPosition);
-      const valueAfterCursor = e.target.value.substring(cursorPosition);
-      
-      // Only add @gmail.com if it doesn't already exist
-      if (!valueBeforeCursor.includes('@gmail.com') && !valueAfterCursor.includes('@gmail.com')) {
-        const newValue = valueBeforeCursor + '@gmail.com' + valueAfterCursor;
-        setInputValue(newValue);
-        
-        const syntheticEvent = {
-          target: {
-            name: e.target.name,
-            value: newValue,
-          }
-        };
-        props.onChange(syntheticEvent);
-        
-        // Set cursor position after @gmail.com
-        setTimeout(() => {
-          e.target.setSelectionRange(cursorPosition + '@gmail.com'.length, cursorPosition + '@gmail.com'.length);
-        }, 0);
-      }
-    }
-  };
-  
-  const handleFocus = (e) => {
-    // When field gets focus and it's empty, add @gmail.com if it's an email field
-    if (enableEmailSuggestions && e.target.value === '') {
-      setInputValue('@gmail.com');
-      
-      const syntheticEvent = {
-        target: {
-          name: e.target.name,
-          value: '@gmail.com',
-        }
-      };
-      
-      props.onChange(syntheticEvent);
-      
-      // Place cursor at beginning of input field
-      setTimeout(() => {
-        e.target.setSelectionRange(0, 0);
-      }, 0);
-    }
-    
-    if (props.onFocus) {
-      props.onFocus(e);
-    }
-  };
-
-  return (
-    <motion.div className="form-group">
-      {(toolTip || props.placeholder) && (
-        <div className={`input-label ${toolTip ? "with-tooltip" : ""}`}>
-          <span>{props.placeholder}</span>
-          {toolTip && (
-            <motion.span
-              className="tooltip-icon"
-              title={toolTip}
-              whileHover={{ scale: 1.2 }}
-            >
-              <FiInfo size={12} />
-            </motion.span>
-          )}
-        </div>
-      )}
-      <div className={`input-field ${error ? "error" : ""}`}>
-        <motion.span
-          className="field-icon"
-          initial={{ opacity: 0.5 }}
-          whileHover={{ opacity: 1 }}
-        >
-          {icon}
-        </motion.span>
-        <input 
-          {...props} 
-          value={inputValue}
-          onChange={enableEmailSuggestions ? handleEmailInput : props.onChange}
-          onKeyDown={enableEmailSuggestions ? handleKeyDown : props.onKeyDown}
-          onFocus={enableEmailSuggestions ? handleFocus : props.onFocus}
-          autoComplete={enableEmailSuggestions ? "off" : props.autoComplete}
-          placeholder={enableEmailSuggestions ? (props.placeholder || "") + " (@gmail.com sẽ được thêm tự động)" : props.placeholder}
-        />
-        {endIcon}
-      </div>
-      
-      {requirements && !error && (
-        <div className="field-requirements">{requirements}</div>
-      )}
-      {error && (
-        <motion.div
-          className="error-message"
-          initial={{ opacity: 0, height: 0 }}
-          animate={{ opacity: 1, height: "auto" }}
-          transition={{ duration: 0.2 }}
-        >
-          {error}
-        </motion.div>
-      )}
-    </motion.div>
-  );
-};
-
-// Email field with fixed domain component
-const EmailField = ({
-  icon,
-  error,
-  toolTip,
-  value,
-  onChange,
-  name,
-  placeholder,
-  ...props
-}) => {
-  // Extract username part from email (before @gmail.com)
-  const [username, setUsername] = useState(
-    value?.includes('@') ? value.split('@')[0] : value || ''
-  );
-
-  // Update parent component when username changes
-  useEffect(() => {
-    // Only extract username if value contains @ (full email format)
-    if (value && value.includes('@')) {
-      setUsername(value.split('@')[0]);
-    }
-  }, [value]);
-
-  // Handle username input changes
-  const handleUsernameChange = (e) => {
-    const newUsername = e.target.value;
-    setUsername(newUsername);
-    
-    // Construct the full email and pass to parent's onChange
-    const fullEmail = newUsername + '@gmail.com';
-    const syntheticEvent = {
-      target: {
-        name: name,
-        value: fullEmail
-      }
-    };
-    
-    onChange(syntheticEvent);
-  };
-
-  return (
-    <motion.div className="form-group">
-      {(toolTip || placeholder) && (
-        <div className={`input-label ${toolTip ? "with-tooltip" : ""}`}>
-          <span>{placeholder}</span>
-          {toolTip && (
-            <motion.span
-              className="tooltip-icon"
-              title={toolTip}
-              whileHover={{ scale: 1.2 }}
-            >
-              <FiInfo size={12} />
-            </motion.span>
-          )}
-        </div>
-      )}
-      <div className={`input-field email-field ${error ? "error" : ""}`}>
-        <motion.span
-          className="field-icon"
-          initial={{ opacity: 0.5 }}
-          whileHover={{ opacity: 1 }}
-        >
-          {icon}
-        </motion.span>
-        <div className="email-input-container">
-          <input
-            type="text"
-            value={username}
-            onChange={handleUsernameChange}
-            placeholder={placeholder ? placeholder.replace("Email", "Tên email") : ""}
-            className="email-prefix"
-            autoComplete="off"
-            {...props}
-          />
-          <div className="email-domain">@gmail.com</div>
-        </div>
-      </div>
-      {error && (
-        <motion.div
-          className="error-message"
-          initial={{ opacity: 0, height: 0 }}
-          animate={{ opacity: 1, height: "auto" }}
-          transition={{ duration: 0.2 }}
-        >
-          {error}
-        </motion.div>
-      )}
-    </motion.div>
-  );
-};
-
-// Forgot Password Form Component
-const ForgotPasswordForm = ({
-  forgotPasswordEmail,
-  setForgotPasswordEmail,
-  handleForgotPassword,
-  onCancel,
-}) => {
-  const [otpSent, setOtpSent] = useState(false);
-  const [loading, setLoading] = useState(false);
-  const [otp, setOtp] = useState('');
-  const [newPassword, setNewPassword] = useState('');
-  const [confirmPassword, setConfirmPassword] = useState('');
-  const navigate = useNavigate();
-
-  const handleSendOtp = async (e) => {
-    e.preventDefault();
-    if (!forgotPasswordEmail || !forgotPasswordEmail.trim()) {
-      message.error("Vui lòng nhập email của bạn!");
-      return;
-    }
-
-    setLoading(true);
-    try {
-      await authApi.sendOtpEmail({ email: forgotPasswordEmail.trim() });
-      message.success("Mã OTP đã được gửi đến email của bạn!");
-      setOtpSent(true);
-    } catch (error) {
-      const errorMessage = error.response?.data?.message || "Không thể gửi OTP. Vui lòng thử lại sau.";
-      message.error(errorMessage);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const handleResetPassword = async (e) => {
-    e.preventDefault();
-    
-    if (!otp || !otp.trim()) {
-      message.error("Vui lòng nhập mã OTP!");
-      return;
-    }
-    
-    if (!newPassword || newPassword.length < 6) {
-      message.error("Mật khẩu mới phải có ít nhất 6 ký tự!");
-      return;
-    }
-    
-    if (newPassword !== confirmPassword) {
-      message.error("Mật khẩu xác nhận không khớp!");
-      return;
-    }
-
-    setLoading(true);
-    try {
-      await authApi.resetPassword({
-        code: otp.trim(),
-        newPassword: newPassword
-      });
-      message.success("Đặt lại mật khẩu thành công! Bạn có thể đăng nhập với mật khẩu mới.");
-      onCancel(); // Return to login form
-    } catch (error) {
-      const errorMessage = error.response?.data?.message || "Đặt lại mật khẩu thất bại. Vui lòng thử lại!";
-      message.error(errorMessage);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  return (
-    <motion.div
-      key="forgot-password"
-      initial={{ opacity: 0, y: 10 }}
-      animate={{ opacity: 1, y: 0 }}
-      exit={{ opacity: 0, y: -10 }}
-      transition={{ duration: 0.3 }}
-      className="forgot-password-form"
-    >
-      {!otpSent ? (
-        <>
-          <p className="form-description">Vui lòng nhập email của bạn để nhận mã OTP đặt lại mật khẩu.</p>
-          <form onSubmit={handleSendOtp} className="password-reset-form">
-            <EmailField
-              icon={<FiMail />}
-              name="email"
-              placeholder="Email của bạn"
-              value={forgotPasswordEmail}
-              onChange={(e) => setForgotPasswordEmail(e.target.value)}
-            />
-            <div className="button-group">
-              <motion.button
-                type="button"
-                className="cancel-button"
-                onClick={onCancel}
-                whileTap={{ scale: 0.98 }}
-              >
-                Hủy
-              </motion.button>
-              <motion.button
-                type="submit"
-                className={`submit-button ${loading ? 'loading' : ''}`}
-                whileTap={{ scale: 0.98 }}
-                disabled={loading}
-              >
-                {loading && <span className="button-spinner"></span>}
-                {loading ? "Đang gửi..." : "Gửi OTP"}
-              </motion.button>
-            </div>
-          </form>
-        </>
-      ) : (
-        <>
-          <p className="form-description">Vui lòng nhập mã OTP đã được gửi đến email của bạn và tạo mật khẩu mới.</p>
-          <form onSubmit={handleResetPassword} className="password-reset-form">
-            <InputField
-              icon={<FiInfo />}
-              type="text"
-              placeholder="Nhập mã OTP"
-              value={otp}
-              onChange={(e) => setOtp(e.target.value)}
-              maxLength={6}
-            />
-            <InputField
-              icon={<FiLock />}
-              type="password"
-              placeholder="Mật khẩu mới"
-              value={newPassword}
-              onChange={(e) => setNewPassword(e.target.value)}
-            />
-            <InputField
-              icon={<FiLock />}
-              type="password"
-              placeholder="Xác nhận mật khẩu mới"
-              value={confirmPassword}
-              onChange={(e) => setConfirmPassword(e.target.value)}
-            />
-            <div className="button-group">
-              <motion.button
-                type="button"
-                className="back-button"
-                onClick={() => setOtpSent(false)}
-                whileTap={{ scale: 0.98 }}
-                disabled={loading}
-              >
-                Quay lại
-              </motion.button>
-              <motion.button
-                type="submit"
-                className={`submit-button ${loading ? 'loading' : ''}`}
-                whileTap={{ scale: 0.98 }}
-                disabled={loading}
-              >
-                {loading && <span className="button-spinner"></span>}
-                {loading ? "Đang xử lý..." : "Đặt lại mật khẩu"}
-              </motion.button>
-            </div>
-          </form>
-        </>
-      )}
-    </motion.div>
-  );
-};
-
-const CustomCheckbox = ({ label, ...props }) => (
-  <label className="custom-checkbox">
-    <input type="checkbox" {...props} />
-    <span className="checkmark" />
-    <span>{label}</span>
-  </label>
-);
-
-const SocialButton = ({ icon, label, color, onClick }) => (
-  <motion.button
-    type="button"
-    className={`social-button ${label.toLowerCase()}`}
-    style={{ "--button-color": color }}
-    onClick={onClick}
-    whileHover={{ scale: 1.05 }}
-    whileTap={{ scale: 0.95 }}
-  >
-    <motion.span className="icon">{icon}</motion.span>
-    <span className="label">{label}</span>
-  </motion.button>
-);
 
 export default LoginForm;
