+++ product-shopthethao/ShopeTheThao_frontend_react/src/components/User/Auth/Login/loginForm.scss
@@ -485,6 +485,7 @@ $border-color: #E2E8F0;
       user-select: none; // Prevent selecting @gmail.com text
       padding-right: 0.5rem;
       background: transparent;
+      margin-left: 5px;
     }
   }
