+++ product-shopthethao/ShopeTheThao_frontend_react/src/components/User/Auth/ResetPassword/ResetPasswordFrom.jsx
@@ -1,0 +1,7 @@
+const ResetPasswordFrom = () => {
+    return (
+  <h1>ssssssss</h1>
+    )
+  }
+  export default ResetPasswordFrom;
+
