+++ product-shopthethao/ShopeTheThao_frontend_react/src/components/User/Auth/index.jsx
@@ -1,0 +1,8 @@
+export { default as RegisterForm } from "./Register/RegisterForm";
+
+
+export { default as ForgotPasswordForm } from "./ForgotPassword/ForgotPasswordForm";
+
+export { default as EmailVerificationForm } from './EmailVerification/EmailVerificationForm';
+
+export * from './Common/FormFields';
