+++ product-shopthethao/ShopeTheThao_frontend_react/src/components/User/Auth/styles/variables.scss
@@ -1,0 +1,24 @@
+// Global variables for auth components
+$primary-color: #4F46E5;
+$secondary-color: #7C3AED;
+$error-color: #dc2626;
+$text-primary: #1E293B;
+$text-secondary: #64748B;
+$border-color: #E2E8F0;
+
+// Common mixins
+@mixin flex-center {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+@mixin glass-effect {
+  background: rgba(255, 255, 255, 0.95);
+  backdrop-filter: blur(10px);
+  -webkit-backdrop-filter: blur(10px);
+}
+
+@mixin smooth-transition {
+  transition: all 0.3s cubic-bezier(0.4, 0, 0.2, 1);
+}
