+++ product-shopthethao/ShopeTheThao_frontend_react/src/pages/User/index.jsx
@@ -2,7 +2,7 @@ export { default as HomeIndex } from "./Home";
 export { default as Products } from "./Products";
 export { default as LoginForm } from "components/User/Auth/Login/LoginForm";
 export { default as OtpForm } from "components/User/Auth/Otp/OtpForm";
-export { default as ChangePasswordForm } from "components/User/Auth/ChangePassword/ChangePasswordForm";
+
 
 
 export { default as Seefulldetails } from "pages/User/Seefulldetails";
